+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-### 0.0.1(2022-07-19)
+### 1.0.3
+
+- add options `guestPath`
+- remove vuex store replace with cookie
+- add function logout()
+- add return value and error log in api refrence
+- add storage class
+
+### 1.0.0
 
 - init